Improved TextDocument and TextRange. Add a simple module system to GUI.

+++ src/CodeSharper.Core/Common/Runnables/CollectionOperations/TakeRunnable.cs
@@ -8,7 +8,7 @@ namespace CodeSharper.Core.Common.Runnables.CollectionRunnables
     public class TakeRunnable : RunnableBase<IEnumerable<Object>, IEnumerable<Object>>
     {
         [Parameter("count")]
-        public Double Count { get; set; }
+        public Int32 Count { get; set; }
 
         public override IEnumerable<Object> Run(IEnumerable<Object> parameter)
         {
@@ -16,7 +16,7 @@ namespace CodeSharper.Core.Common.Runnables.CollectionRunnables
             {
                 return Enumerable.Empty<Object>();
             }
-            return parameter.ToArray().Take((Int32) Count);
+            return parameter.ToArray().Take(Count);
         }
     }
 }
